Fix bug #326 where GeoProcessor crashed at startup loading a DLL.
This was due to the manual copy of DLLs to Python environment to
make sure pip worked.
The fix is to remove those DLLs when done with pip.
The run batch files were cleaned up a lot based on troubleshooting,
including adding new options to select Python, virtual envrironment
tool, level of output, etc.